+++ otterdog/models/__init__.py
@@ -199,6 +199,9 @@ class ModelObject(ABC):
             to_value = self.__getattribute__(key)
             from_value = other.__getattribute__(key)
 
+            if not other.include_field_for_diff_computation(other._get_field(key)):
+                continue
+
             if is_unset(from_value):
                 continue
 
@@ -220,6 +223,9 @@ class ModelObject(ABC):
         ):
             value = self.__getattribute__(key)
             other_value = other.__getattribute__(key)
+
+            if not other.include_field_for_patch_computation(other._get_field(key)):
+                continue
 
             if is_unset(other_value):
                 continue