+++ otterdog/models/webhook.py
@@ -61,7 +61,7 @@ class Webhook(ModelObject, abc.ABC):
     def include_field_for_diff_computation(self, field: dataclasses.Field) -> bool:
         match field.name:
             case "secret":
-                return False
+                return not self.has_dummy_secret()
             case _:
                 return True