+++ otterdog/operations/diff_operation.py
@@ -174,7 +174,7 @@ class DiffOperation(Operation):
         expected_org.generate_live_patch(current_org, context, handle)
 
         # add a warning that otterdog potentially must be run a second time
-        # to fully apply all setting.
+        # to fully apply all settings.
         if "web_commit_signoff_required" in context.modified_org_settings:
             change = context.modified_org_settings["web_commit_signoff_required"]
             if change.to_value is False: